functionsについてまとめた
